# shared/contracts/api-contract.yml
openapi: 3.1.0
info:
  title: USORA KYC Platform API
  version: 1.0.0
  description: >
    Enterprise-grade, multi-tenant Know Your Customer (KYC) platform API.
    Supports identity verification workflows including document validation,
    biometric matching, risk scoring, and compliance screening across
    multiple tenants with isolated configuration and data.
servers:
  - url: https://api.usora.io
    description: Production server
  - url: https://staging.api.usora.io
    description: Staging server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
  - mTLS: []
paths:
...
  /v1/{tenantId}/cases:
    get:
      operationId: listCases
      summary: List cases for a tenant
      tags:
        - Cases
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: page_token
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: assigned_to
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of cases
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Case'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
...
  /v1/{tenantId}/cases/{caseId}/status:
    put:
      operationId: updateCaseStatus
      summary: Update case status
      tags:
        - Cases
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaseStatusUpdate'
      responses:
        '200':
          description: Case updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Case'
        '400':
          description: Invalid status update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/{tenantId}/config:
    get:
      operationId: getTenantConfig
      summary: Get tenant configuration
      tags:
        - Tenant Configuration
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Tenant configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantConfig'
        '404':
          description: Tenant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      operationId: updateTenantConfig
      summary: Update tenant configuration
      tags:
        - Tenant Configuration
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TenantConfig'
      responses:
        '200':
          description: Configuration updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantConfig'
        '400':
          description: Invalid configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained during tenant authentication
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: Tenant API key for server-to-server communication
    mTLS:
      type: mutualTLS
      description: Mutual TLS for internal service-to-service communication
  schemas:
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Application-specific error code
          example: VERIFICATION_NOT_FOUND
        message:
          type: string
          description: Human-readable error message
          example: The requested verification was not found
        target:
          type: string
          description: The target of the error (field or resource)
          example: verificationId
        details:
          type: array
          items:
            type: object
            properties:
              code:
                type: string
              message:
                type: string
              target:
                type: string
        request_id:
          type: string
          format: uuid
          description: Unique identifier for the request
        timestamp:
          type: string
          format: date-time
          description: When the error occurred
        documentation_url:
          type: string
          format: uri
          description: Link to documentation for this error
          example: https://docs.usora.io/errors/VERIFICATION_NOT_FOUND
...
    Pagination:
      type: object
      properties:
        page_size:
          type: integer
          description: Number of items per page
        page_token:
          type: string
          description: Token for the current page
        next_page_token:
          type: string
          description: Token for the next page (empty if no more results)
        total_count:
          type: integer
          description: Total number of items matching the query
...
    Case:
      type: object
      required:
        - id
        - tenant_id
        - status
        - created_at
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        verification_id:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - open
            - under_review
            - pending_info
            - resolved_approved
            - resolved_rejected
            - escalated
        assigned_to:
          type: string
          description: Reviewer identifier
        created_at:
          type: string
          format: date-time
        resolved_at:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
    CaseStatusUpdate:
      type: object
      properties:
        status:
          type: string
          enum:
            - open
            - under_review
            - pending_info
            - resolved_approved
            - resolved_rejected
            - escalated
        assigned_to:
          type: string
        notes:
          type: string
        reason:
          type: string
          description: Reason for the status change
    TenantConfig:
      type: object
      required:
        - tenant_id
      properties:
        tenant_id:
          type: string
          format: uuid
        features:
          type: array
          items:
            type: string
          description: Enabled feature flags for the tenant
        rate_limits:
          type: object
          properties:
            requests_per_second:
              type: integer
            burst_size:
              type: integer
            concurrent_verifications:
              type: integer
        webhook_urls:
          type: array
          items:
            type: string
            format: uri
        allowed_document_types:
          type: array
          items:
            type: string
        kyc_config:
          type: object
          properties:
            document_verification:
              type: boolean
            biometric_matching:
              type: boolean
            liveness_check:
              type: boolean
            risk_scoring:
              type: boolean
            aml_screening:
              type: boolean
            min_confidence_score:
              type: number
              minimum: 0
              maximum: 1
        branding:
          type: object
          properties:
            primary_color:
              type: string
            logo_url:
              type: string
              format: uri
            company_name:
              type: string
...